+++ index.ts
@@ -1,1 +1,2 @@
 export * from './src/wobbly'
+export * from './src/spawn'